Third Year/First Semester/Python/GUI-Resource-Monitor/requirements.txt
matplotlib==3.10.8
psutil==7.2.0
py-cpuinfo==9.0.0
